feat: Move chat storing logic to separate file
